requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
PyJWT==2.3.0
python-dotenv==0.19.1
openai==0.27.0
gunicorn==20.1.0
psycopg2-binary==2.9.1